requirements.txt
Flask==3.0.2
pymongo==4.6.1
python-dotenv==1.0.1
marshmallow==3.20.2
PyJWT==2.8.0
Flask-JWT-Extended==4.6.0
Werkzeug==3.0.3
